KLUDGE! declare BPR for new library cards that have BP.
this is a stop-gap solution until the cards API endpoint can handle the
new bloodPotency property for library cards, and the client-side card
decoder, models, and UI have been updated accordingly.
for practical purposes, both can be treated synonymously for now, as
there are no cards that have BP and BPR at the same time. the
client-side card filter in the deckbuilder currently makes no difference
between the two, as BPR on library cards is mapped to BP there anyway.